apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.133.0",
    "uvicorn>=0.41.0",
    "pydantic-settings>=2.7.1",
    "supabase>=2.13.0",
    "upstash-redis>=1.3.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1",
    "google-genai>=0.5.0",
]
